v2

+++ migrations/006.do.create_progress.sql
@@ -4,7 +4,7 @@ CREATE TABLE progress (
     date_modified TIMESTAMP,
     book_id INTEGER REFERENCES books(id) ON DELETE CASCADE NOT NULL,
     user_id INTEGER REFERENCES users(id) ON DELETE CASCADE NOT NULL,
-    percent INTEGER DEFAULT 0 NOT NULL,
+    percent DECIMAL DEFAULT 0 NOT NULL,
     reading_status TEXT NOT NULL,
     pagecount INTEGER NOT NULL,
     maxpagecount INTEGER NOT NULL
